[3726] Convert CalculatorPresenters to a concern
